switched to true css color vars, renamed component dir
