+++ components/base.css
@@ -1,0 +1,12 @@
+:root {
+    --color-orange: #ff694b; /* "dbt-orange" */
+    --color-teal-500: #38b2ac;	
+    --color-teal-700: #2c7a7b;
+    --color-white: #fff;
+    --color-transparent: transparent;	
+    --color-red: #e21b10;
+    --color-gray-400: #cbd5e0;
+    --color-gray-500: #a0aec0;
+    --color-gray-lighter: #f9fafb;
+    --color-gray-dark: #a9a9a9;
+}
