+++ components/custom-components.css
@@ -1,3 +1,4 @@
+@import './base.css';
 @import './Button/button.css';
 @import './Checkbox/checkbox.css';
-@import './Input/inputListItem.css';
+@import './InputList/inputListItem.css';
