+++ tailwind.config.js
@@ -1,20 +1,28 @@
 module.exports = {
   prefix: "tw-",
   theme: {
-    extend: {
-      colors: {
-        gray: {
-          lighter: "#f9fafb",
-          dark: "#a9a9a9",
-          400: "#cbd5e0",
-          500: "#a0aec0"
-        },
-        orange: {
-          default: "#ff694b", // "dbt-orange"
-        },
-        red: {
-          default: "#e21b10",
-        },
+    colors: {
+      transparent: {
+        default: 'var(--color-transparent)',
+      },
+      white: {
+        default: 'var(--color-white)',
+      },
+      teal: {
+        500: 'var(--color-teal-500)',
+        700: 'var(--color-teal-700)',
+      },
+      orange: {
+        default: 'var(--color-orange)',
+      },
+      gray: {
+        lighter:'var(--color-gray-lighter)',
+        400: 'var(--color-gray-400)',
+        500: 'var(--color-gray-500)',
+        dark: 'var(--color-gray-dark)',
+      },
+      red: {
+        default:'var(--color-red)',
       },
     },
   },
